+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+* The ability to specify a certificate and key with `--cert` and `--key` respectively has been replaced with the `--certs` option.
+  Certificates are now stored in a special directory. To migrate to this version, the keys should be stored in the `.certificates` directory (or any other directory you specify).
+  This enables us to use multiple certificates for multiple domains.
+
 ### Fixed
+* Agate now requires the use of SNI by any connecting client.
 * All log lines are in the same format now:
   `<local ip>:<local port> <remote ip or dash> "<request>" <response status> "<response meta>" [error:<error>]`
   If the connection could not be established correctly (e.g. because of TLS errors), the status code `00` is used.